chore: refactor Dialog components into folders
